backend/requirements.txt
fastapi
uvicorn
requests
sentence-transformers
faiss-cpu
python-dotenv
google-genai
